Pull out basecoin stuff
